split generator/main.go
